+++ src/plugin/context/systemparams/rapier_context_systemparam.rs
@@ -4,7 +4,7 @@ use bevy::prelude::*;
 use rapier::prelude::Real;
 
 pub(crate) const RAPIER_CONTEXT_EXPECT_ERROR: &str =
-    "RapierContextEntityLink.0 refers to an entity missing components from RapierContextBundle.";
+    "RapierContextEntityLink.0 refers to an entity missing components from RapierContextSimulation.";
 
 use crate::{
     plugin::context::{
@@ -13,9 +13,6 @@ use crate::{
     },
     prelude::QueryFilter,
 };
-
-#[cfg(doc)]
-use crate::prelude::RapierContextBundle;
 
 /// Utility [`SystemParam`] to easily access every required components of a [`RapierContext`] immutably.
 ///
@@ -59,7 +56,7 @@ impl<'w, 's, T: query::QueryFilter + 'static> ReadRapierContext<'w, 's, T> {
 /// A helper struct to avoid passing too many parameters to most rapier functions.
 /// This helps with reducing boilerplate, at the (small) price of maybe getting too much information from the ECS.
 ///
-/// Note: This is not a component, refer to [`ReadRapierContext`], [`WriteRapierContext`], or [`RapierContextBundle`]
+/// Note: This is not a component, refer to [`ReadRapierContext`], [`WriteRapierContext`], or [`RapierContextSimulation`]
 #[derive(query::QueryData)]
 pub struct RapierContext<'a> {
     /// The Rapier context, containing all the state of the physics engine.