+++ src/plugin/systems/multiple_rapier_contexts.rs
@@ -130,13 +130,9 @@ fn bubble_down_context_change(
 
 #[cfg(test)]
 mod test {
-    use crate::plugin::context::RapierQueryPipeline;
     use crate::plugin::systems::tests::HeadlessRenderPlugin;
     use crate::plugin::{
-        context::{
-            RapierContextColliders, RapierContextEntityLink, RapierContextJoints,
-            RapierContextSimulation, RapierRigidBodySet,
-        },
+        context::{RapierContextEntityLink, RapierContextSimulation},
         NoUserData, PhysicsSet, RapierPhysicsPlugin,
     };
     use crate::prelude::{ActiveEvents, Collider, ContactForceEventThreshold, RigidBody, Sensor};
@@ -173,15 +169,7 @@ mod test {
                 .unwrap_or_else(|| panic!("no link to rapier context entity from {entity}."));
         }
         // Verify link is correctly updated for children.
-        let new_rapier_context = world
-            .spawn((
-                RapierContextSimulation::default(),
-                RapierContextColliders::default(),
-                RapierContextJoints::default(),
-                RapierQueryPipeline::default(),
-                RapierRigidBodySet::default(),
-            ))
-            .id();
+        let new_rapier_context = world.spawn((RapierContextSimulation::default(),)).id();
         // FIXME: We need to wait 1 frame when creating a context.
         // Ideally we should be able to order the systems so that we don't have to wait.
         app.update();