+++ CHANGELOG.md
@@ -21,12 +21,12 @@
 - Update from rapier `0.21` to rapier `0.22`,
   see [rapier's changelog](https://github.com/dimforge/rapier/blob/master/CHANGELOG.md).
 - Update bevy to 0.15.
-- `RapierContext`, `RapierConfiguration` and `RenderToSimulationTime` are now a `Component` instead of resources.
+- `RapierContext`, `RapierConfiguration` and `SimulationToRenderTime` are now a `Component` instead of resources.
   - Rapier now supports multiple independent physics worlds, see example `multi_world3` for usage details.
   - Migration guide:
     - `ResMut<mut RapierContext>` -> `WriteDefaultRapierContext`
     - `Res<RapierContext>` -> `ReadDefaultRapierContext`
-    - Access to `RapierConfiguration` and `RenderToSimulationTime` should query for it
+    - Access to `RapierConfiguration` and `SimulationToRenderTime` should query for it
 on the responsible entity owning the `RenderContext`.
   - If you are building a library on top of `bevy_rapier` and would want to support multiple independent physics worlds too,
 you can check out the details of [#545](https://github.com/dimforge/bevy_rapier/pull/545)